+++ pkg/api/group.go
@@ -6,7 +6,7 @@ import (
 )
 
 // 群组踢人
-func SetGroupKick(groupId, userId int, rejectAddRequest bool) error {
+func SetGroupKick(groupId, userId int64, rejectAddRequest bool) error {
 	req := make(map[string]interface{})
 	req["user_id"] = userId
 	req["group_id"] = groupId
@@ -21,8 +21,8 @@ func SetGroupKick(groupId, userId int, rejectAddRequest bool) error {
 }
 
 // 群组禁言
-func SetGroupBan(groupId, userId, duration int) error {
-	req := make(map[string]int)
+func SetGroupBan(groupId, userId, duration int64) error {
+	req := make(map[string]int64)
 	req["user_id"] = userId
 	req["group_id"] = groupId
 	req["duration"] = duration
@@ -35,7 +35,7 @@ func SetGroupBan(groupId, userId, duration int) error {
 }
 
 // 群组匿名用户禁言
-func SetGroupAnonymousBan(groupId, duration int, anonymousFlag string) error {
+func SetGroupAnonymousBan(groupId, duration int64, anonymousFlag string) error {
 	req := make(map[string]interface{})
 	req["group_id"] = groupId
 	req["anonymous_flag"] = anonymousFlag
@@ -48,7 +48,7 @@ func SetGroupAnonymousBan(groupId, duration int, anonymousFlag string) error {
 }
 
 //群组全员禁言
-func SetGroupWholeBan(groupId int, enable bool) error {
+func SetGroupWholeBan(groupId int64, enable bool) error {
 	req := make(map[string]interface{})
 	req["group_id"] = groupId
 	req["enable"] = enable
@@ -62,7 +62,7 @@ func SetGroupWholeBan(groupId int, enable bool) error {
 }
 
 //群组设置管理员
-func SetGroupAdmin(groupId, userId int, enable bool) error {
+func SetGroupAdmin(groupId, userId int64, enable bool) error {
 	req := make(map[string]interface{})
 	req["group_id"] = groupId
 	req["user_id"] = userId
@@ -77,7 +77,7 @@ func SetGroupAdmin(groupId, userId int, enable bool) error {
 }
 
 //群组匿名
-func SetGroupAnonymous(groupId int, enable bool) error {
+func SetGroupAnonymous(groupId int64, enable bool) error {
 	req := make(map[string]interface{})
 	req["group_id"] = groupId
 	req["enable"] = enable
@@ -90,7 +90,7 @@ func SetGroupAnonymous(groupId int, enable bool) error {
 }
 
 //设置群名片（群备注）
-func SetGroupCard(groupId, userId int, card string) error {
+func SetGroupCard(groupId, userId int64, card string) error {
 	req := make(map[string]interface{})
 	req["group_id"] = groupId
 	req["user_id"] = userId
@@ -104,7 +104,7 @@ func SetGroupCard(groupId, userId int, card string) error {
 }
 
 //设置群名
-func SetGroupName(groupId int, groupName string) error {
+func SetGroupName(groupId int64, groupName string) error {
 	req := make(map[string]interface{})
 	req["group_id"] = groupId
 	req["group_name"] = groupName
@@ -117,7 +117,7 @@ func SetGroupName(groupId int, groupName string) error {
 }
 
 //退出群组
-func SetGroupLeave(groupId int, isDismiss bool) error {
+func SetGroupLeave(groupId int64, isDismiss bool) error {
 	req := make(map[string]interface{})
 	req["group_id"] = groupId
 	req["is_dismiss"] = isDismiss
@@ -130,7 +130,7 @@ func SetGroupLeave(groupId int, isDismiss bool) error {
 }
 
 //设置群组专属头衔
-func SetGroupSpecialTitle(groupId, userId, duration int, specialTitle string) error {
+func SetGroupSpecialTitle(groupId, userId, duration int64, specialTitle string) error {
 	req := make(map[string]interface{})
 	req["group_id"] = groupId
 	req["user_id"] = userId
@@ -160,7 +160,7 @@ func SetGroupAddRequest(flag, subType, reason string, approve bool) error {
 }
 
 //获取群信息
-func GetGroupInfo(groupId int, noCache bool) (*model.Group, error) {
+func GetGroupInfo(groupId int64, noCache bool) (*model.Group, error) {
 	req := make(map[string]interface{})
 	req["group_id"] = groupId
 	req["no_cache"] = noCache
@@ -187,7 +187,7 @@ func GetGroupList() ([]*model.Group, error) {
 }
 
 //获取群成员信息
-func GetGroupMemberInfo(groupId, userId int, noCache bool) (*model.GroupMember, error) {
+func GetGroupMemberInfo(groupId, userId int64, noCache bool) (*model.GroupMember, error) {
 	req := make(map[string]interface{})
 	req["group_id"] = groupId
 	req["user_id"] = userId
@@ -213,7 +213,7 @@ func GetGroupMemberListInfo() ([]*model.GroupMember, error) {
 }
 
 //获取群荣誉信息
-func GetGroupHonorInfo(groupId int, honorType string) (*model.GroupHonorInfo, error) {
+func GetGroupHonorInfo(groupId int64, honorType string) (*model.GroupHonorInfo, error) {
 	req := make(map[string]interface{})
 	req["group_id"] = groupId
 	req["type"] = honorType
